refactor(jsonld): move json ld normalisation and digestion utils to /linkedData folder
